Remove browser module
